requirements.txt
annotated-types==0.7.0
anyio==4.8.0
audioread==3.0.1
boto3==1.36.6
botocore==1.36.6
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
decorator==5.1.1
fastapi==0.115.7
h11==0.14.0
httptools==0.6.4
idna==3.10
jmespath==1.0.1
joblib==1.4.2
lazy_loader==0.4
librosa==0.10.2.post1
llvmlite==0.44.0
msgpack==1.1.0
numba==0.61.0
numpy==2.1.3
packaging==24.2
platformdirs==4.3.6
pooch==1.8.2
py-youtube==1.1.7
pycparser==2.22
pydantic==2.10.5
pydantic_core==2.27.2
pydub==0.25.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
requests==2.32.3
s3transfer==0.11.2
scikit-learn==1.6.1
scipy==1.15.1
six==1.17.0
sniffio==1.3.1
soundfile==0.13.1
soxr==0.5.0.post1
starlette==0.45.2
threadpoolctl==3.5.0
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
watchfiles==1.0.4
websockets==14.2
yt-dlp==2025.1.15
